refactor: unify in-memory cache to simplify usage
