+++ foyer-memory/src/prelude.rs
@@ -12,44 +12,10 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-use ahash::RandomState;
-
-use crate::{
-    cache::{Cache, CacheConfig, CacheEntry},
-    eviction::{
-        fifo::{Fifo, FifoHandle},
-        lfu::{Lfu, LfuHandle},
-        lru::{Lru, LruHandle},
-    },
-    indexer::HashTableIndexer,
-    listener::DefaultCacheEventListener,
-};
-
-pub type FifoCache<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> =
-    Cache<K, V, FifoHandle<K, V>, Fifo<K, V>, HashTableIndexer<K, FifoHandle<K, V>>, L, S>;
-pub type FifoCacheConfig<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> = CacheConfig<Fifo<K, V>, L, S>;
-pub type FifoCacheEntry<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> =
-    CacheEntry<K, V, FifoHandle<K, V>, Fifo<K, V>, HashTableIndexer<K, FifoHandle<K, V>>, L, S>;
-
-pub type LruCache<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> =
-    Cache<K, V, LruHandle<K, V>, Lru<K, V>, HashTableIndexer<K, LruHandle<K, V>>, L, S>;
-pub type LruCacheConfig<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> = CacheConfig<Lru<K, V>, L, S>;
-pub type LruCacheEntry<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> =
-    CacheEntry<K, V, LruHandle<K, V>, Lru<K, V>, HashTableIndexer<K, LruHandle<K, V>>, L, S>;
-
-pub type LfuCache<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> =
-    Cache<K, V, LfuHandle<K, V>, Lfu<K, V>, HashTableIndexer<K, LfuHandle<K, V>>, L, S>;
-pub type LfuCacheConfig<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> = CacheConfig<Lfu<K, V>, L, S>;
-pub type LfuCacheEntry<K, V, L = DefaultCacheEventListener<K, V>, S = RandomState> =
-    CacheEntry<K, V, LfuHandle<K, V>, Lfu<K, V>, HashTableIndexer<K, LfuHandle<K, V>>, L, S>;
-
 pub use crate::{
-    cache::Entry,
+    cache::{Cache, CacheEntry, Entry},
     context::CacheContext,
-    eviction::{
-        fifo::{FifoConfig, FifoContext},
-        lfu::{LfuConfig, LfuContext},
-        lru::{LruConfig, LruContext},
-    },
+    eviction::{fifo::FifoConfig, lfu::LfuConfig, lru::LruConfig},
     listener::CacheEventListener,
+    metrics::Metrics,
 };
